requirements.txt
langchain-community
langchain-openai
openai
python-dotenv
langserve[all]
langchain-chroma
langchain-openai
faiss-cpu
langgraph